dom/media/test/test_mediarecorder_mp4_support.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Media Recording - test mp4 MIME support</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<pre id="test">
<script class="testbody" type="text/javascript">

ok(MediaRecorder.isTypeSupported('video/mp4'), 'Should support video/mp4');

</script>
</head>
</html>
